+++ documentation/ChangeLog.md
@@ -11,6 +11,7 @@
  * Added build console preference page. 
   * Can now customize build console text colors.
   * Activate console on error message option.
+ * Added #45 - Allow specifying extra options to 'go build' (available in project properties). They default to `-gcflags "-N -l"`
 
 
 ### Goclipse 0.8.0